Fix msLoadingIndicatorCustom.cpp for an updated VC++ compiler.

+++ src/mvLoadingIndicatorCustom.cpp
@@ -97,7 +97,7 @@ void LoadingIndicatorCircle2(const char* label, float indicatorRadiusFactor, flo
 
     int num_segments = 30;
 
-    int start = abs(ImSin(g.Time * 1.8f) * (num_segments - 5));
+    int start = abs(int(ImSin(g.Time * 1.8f) * (num_segments - 5)));
 
     const float a_min = IM_PI * 2.0f * ((float)start) / (float)num_segments;
     const float a_max = IM_PI * 2.0f * ((float)num_segments - 3) / (float)num_segments;
